backend/requirements.txt
django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0
requests==2.31.0
pypdf2==3.0.1
numpy>=1.24.0
langchain>=0.0.267
langchain-community>=0.0.10
drf-yasg==1.21.7

# Optional dependencies
# pandas>=2.0.0